0.0.56 Pre-alpha - Completion of API localization; - Completion of localization of the server side of web pages; - Completion of localization of the client side of web pages; - Adding a method to the kernel localization class to retrieve a single value from the localization file for further translation; - Minor fixes;

+++ api/user/delete.handler.php
@@ -15,24 +15,32 @@ if (!defined('IS_NOT_HACKED')) {
 
 use \core\PHPLibrary\User as User;
 
-if (isset($_DELETE['user_id'])) {
-  $user_id = (is_numeric($_DELETE['user_id'])) ? (int)$_DELETE['user_id'] : 0;
+if ($system_core->client->is_logged(2)) {
+  if (isset($_DELETE['user_id'])) {
+    $user_id = (is_numeric($_DELETE['user_id'])) ? (int)$_DELETE['user_id'] : 0;
 
-  if (User::exists_by_id($system_core, $user_id)) {
-    $user = new User($system_core, $user_id);
+    if (User::exists_by_id($system_core, $user_id)) {
+      $user = new User($system_core, $user_id);
 
-    $user_is_deleted = $user->delete();
-    if ($user_is_deleted) {
-      $handler_message = (!isset($handler_message)) ? 'Пользователь успешно удален.' : $handler_message;
-      $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+      $user_is_deleted = $user->delete();
+      if ($user_is_deleted) {
+        $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS') : $handler_message;
+        $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+      } else {
+        $handler_message = (!isset($handler_message)) ? (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+        $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+      }
     } else {
-      $handler_message = (!isset($handler_message)) ? 'Пользователь не был удален, поскольку произошел неизвестный сбой.' : $handler_message;
+      $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_USER_ERROR_NOT_FOUND')) : $handler_message;
       $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
     }
   } else {
-    $handler_message = (!isset($handler_message)) ? 'Данные пользователя не были сохранены, поскольку его не существует.' : $handler_message;
+    $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_INVALID_INPUT_DATA_SET')) : $handler_message;
     $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
   }
+} else {
+  $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_AUTHORIZATION')) : $handler_message;
+  $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
 }
 
 ?>
